appstore/pyproject.toml
[project]
name = "appstore"
version = "0.1.0"
description = ""
authors = [
    {name = "Jerenemy",email = "jzay@wesleyan.edu"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=12.1.1,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
